+++ play/sonolus-cli.config.mjs
@@ -1,6 +1,5 @@
 import { error, log } from 'node:console'
 import { copyFileSync, readFileSync } from 'node:fs'
-import { resolve } from 'node:path'
 import { hash } from 'sonolus-core'
 
 /** @type {import('sonolus.js').SonolusCLIConfig} */
@@ -20,35 +19,6 @@ export default {
         } catch (_) {
             error('Error: failed to setup bgm, using fallback')
             log()
-        }
-    },
-    async esbuild(options) {
-        return {
-            ...options,
-            plugins: [
-                {
-                    name: 'ts-paths',
-                    setup(build) {
-                        build.onResolve(
-                            {
-                                filter: /^~(lib)?\/.*/,
-                            },
-                            async (args) => {
-                                const path = args.path.startsWith('~lib/')
-                                    ? './lib/' + args.path.slice(5).replace('.cjs', '.cts')
-                                    : args.path.startsWith('~shared/')
-                                    ? './shared/src/' + args.path.slice(8).replace('.mjs', '.mts')
-                                    : './play/src/' + args.path.slice(2).replace('.mjs', '.mts')
-
-                                return {
-                                    path: resolve(process.cwd(), path),
-                                }
-                            }
-                        )
-                    },
-                },
-                ...options.plugins,
-            ],
         }
     },
 }
