+++ play/src/engine/playData/archetypes/InputManager.mts
@@ -1,4 +1,3 @@
-import { options } from '../../configuration/options.mjs'
 import { ClaimManager } from './ClaimManager.mjs'
 
 const disallowedEmpties = levelMemory({
@@ -67,8 +66,6 @@ export class InputManager extends Archetype {
     }
 
     updateSequential() {
-        if (options.autoplay) return
-
         claimStartManager.clear()
 
         claimEndManager.clear()
@@ -81,8 +78,6 @@ export class InputManager extends Archetype {
     }
 
     touch() {
-        if (options.autoplay) return
-
         disallowedTraceStarts.clear()
 
         for (const touch of touches) {
