+++ play/src/engine/playData/archetypes/SimLine.mts
@@ -98,15 +98,11 @@ export class SimLine extends Archetype {
     }
 
     updateParallel() {
-        if (options.autoplay) {
-            if (time.now >= this.leftTargetTime) this.despawn = true
-        } else {
-            if (
-                this.aInfo.state === EntityState.Despawned ||
-                this.bInfo.state === EntityState.Despawned
-            )
-                this.despawn = true
-        }
+        if (
+            this.aInfo.state === EntityState.Despawned ||
+            this.bInfo.state === EntityState.Despawned
+        )
+            this.despawn = true
         if (this.despawn) return
 
         if (options.hidden > 0 && time.now > this.leftVisualTime.hidden) return
