+++ play/src/engine/playData/archetypes/notes/flatNotes/FlatNote.mts
@@ -122,13 +122,7 @@ export abstract class FlatNote extends Note {
 
         this.z = getZ(layer.note.body, this.targetTime, this.data.lane)
 
-        if (options.autoplay) {
-            this.result.judgment = Judgment.Perfect
-
-            this.result.bucket.index = this.bucket.index
-        } else {
-            this.result.accuracy = this.windows.good.max
-        }
+        this.result.accuracy = this.windows.good.max
     }
 
     setLayout({ l, r }: { l: number; r: number }) {
@@ -148,7 +142,6 @@ export abstract class FlatNote extends Note {
     }
 
     updateParallel() {
-        if (options.autoplay && time.now >= this.targetTime) this.despawn = true
         if (time.now > this.inputTime.max) this.despawn = true
         if (this.despawn) return
 
@@ -163,19 +156,17 @@ export abstract class FlatNote extends Note {
     }
 
     terminate() {
-        if (!options.autoplay) return
-
         if (options.noteEffectEnabled) this.playNoteEffects()
         if (options.slotEffectEnabled) this.playSlotEffects(this.targetTime)
         if (options.laneEffectEnabled) this.playLaneEffects()
     }
 
     get shouldScheduleSFX() {
-        return options.sfxEnabled && (options.autoplay || options.autoSFX)
+        return options.sfxEnabled && options.autoSFX
     }
 
     get shouldPlaySFX() {
-        return options.sfxEnabled && !options.autoplay && !options.autoSFX
+        return options.sfxEnabled && !options.autoSFX
     }
 
     get useFallbackSprites() {
