+++ play/src/engine/playData/archetypes/notes/flatNotes/tapNotes/TapNote.mts
@@ -1,4 +1,3 @@
-import { options } from '../../../../../configuration/options.mjs'
 import { claimStart, disallowEmpty, disallowEnd, getClaimedStart } from '../../../InputManager.mjs'
 import { windows } from '../../../windows.mjs'
 import { FlatNote } from '../FlatNote.mjs'
@@ -7,16 +6,12 @@ export abstract class TapNote extends FlatNote {
     leniency = 0.75
 
     updateSequential() {
-        if (options.autoplay) return
-
         if (time.now < this.inputTime.min) return
 
         claimStart(this.info.index, this.targetTime, this.hitbox, this.fullHitbox)
     }
 
     touch() {
-        if (options.autoplay) return
-
         if (time.now < this.inputTime.min) return
 
         const index = getClaimedStart(this.info.index)
