+++ play/tsconfig.json
@@ -1,6 +1,8 @@
 {
     "compilerOptions": {
-        "lib": ["ESNext"],
+        "lib": [
+            "ESNext"
+        ],
         "target": "ESNext",
         "module": "Node16",
         "moduleResolution": "Node16",
@@ -9,11 +11,10 @@
         "skipLibCheck": true,
         "exactOptionalPropertyTypes": true,
         "forceConsistentCasingInFileNames": true,
-        "rootDir": "../",
         "paths": {
             "~/*": ["./src/*"],
-            "~lib/*": ["./lib/*"],
-            "~shared/*": ["./shared/*"]
+            "~lib/*": ["../lib/*"],
+            "~shared/*": ["../shared/src/*"]
         }
     }
 }
